requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langgraph
httpx
python-telegram-bot[aio]
python-dotenv
langchain-community
langchain-core
watchdog
pandas
openpyxl
xlrd
python-multipart
